+++ admin/deps.php
@@ -53,7 +53,7 @@
 										<td><?php echo $dep['name']; ?></td>
 										 <td><?php echo $deps->get_deps_meta( $dep['id'], 'thumbnail', false ) ?></td>
 										  <td>
-											<a href="modules/admin_delete_deps.php?id=<?php echo $dep['id'] ?>"> Delete </a>
+											<a  onclick="return confirm('Do you really want to delete the department?')" href="modules/admin_delete_deps.php?id=<?php echo $dep['id'] ?>"> Delete </a>
 										  </td>
 										   </tr>
 										 <tr>
